Initial commit of `preprint+`.
This is a fork of the [`preprint-template.tex`](https://github.com/brenhinkeller/preprint-template.tex/blob/master/preprint.sty) (itself a fork of the [`arxiv-style`](https://github.com/kourgeorge/arxiv-style) template) with improved support for manuscripts that will be submitted to journals. This lead to the following user-facing changes:

- A rich collection of option flags making every feature optional.
  This allows authors to turn off features that may clash with the journal’s template.
- Additional options to set fonts, title style and page dimensions.
  + This is not an all-you-can eat template.
    The idea is to provide a few useful sets of configurations,
    so that authors are not locked in to using Times New Roman.
- A revised interface for specifying abstract and keywords in the preamble, and have them typeset as part of `\maketitle`. This is more in line with how one would prepare a manuscript using e.g. Springer-Nature’s or PNAS’ templates.
- Added a `\supplementary` command for typesetting Supplementary material as part of the PDF. This allows cross-references between the main text and supplementary, while still making it easy to submit the latter to the journal as a separate file.
  This is distinct from the `\appendix`, which is usually typeset as part of the text, just after references.

Under the hood, the style file has been completely rewritten in LaTeX3 syntax, which is how we are able to provide all of these options without it becoming an unmaintainable mess.